HHH-13432 - EntityManagerFactory no longer exposes "javax.persistence.nonJtaDataSource"
